Combine all test code into one file
