+++ monument/test/runner.rs
@@ -1,5 +1,4 @@
-//! Code that's common between the test and benchmark runners (`test.rs` and `bench.rs`,
-//! respectively).
+//! Runner for test cases (not benchmarks)
 
 use std::{
     collections::{BTreeMap, HashMap, HashSet},
@@ -28,6 +27,53 @@ const TEST_DIRS: [(&str, CaseState); 2] = [
     ("examples/", CaseState::Parse), // Examples to show how Monument's TOML format works
 ];
 const PATH_TO_MONUMENT_DIR: &str = "../";
+
+///////////////////
+// MAIN FUNCTION //
+///////////////////
+
+/// Collect and run all the test cases, but no benchmarks (which would take too long)
+fn main() -> anyhow::Result<()> {
+    // Read the args to determine what to do
+    let mut args = std::env::args();
+    args.next(); // Skip the first arg, which always the path to the test's binary
+    if args.next().as_deref() == Some("bless") {
+        match args.next().as_deref() {
+            // `cargo bless`: bless only unspecified
+            None => bless_tests(BlessLevel::OnlyUnspecified),
+            // `cargo bless --fails ...`
+            Some("--fails") => match args.next() {
+                // `cargo bless --fails {something} ...`
+                Some(arg) => {
+                    println!("Unknown additional arg '{}'", arg);
+                    print_bless_usage();
+                    Ok(())
+                }
+                // `cargo bless --fails`: bless everything
+                None => bless_tests(BlessLevel::Fails),
+            },
+            // `cargo bless {something not --fails} ...`
+            Some(arg) => {
+                println!("Unknown arg '{}'", arg);
+                print_bless_usage();
+                Ok(())
+            }
+        }
+    } else {
+        // If no args were given, just run the tests
+        match run()? {
+            Outcome::Fail => Err(anyhow::Error::msg("Tests failed")),
+            Outcome::Pass => Ok(()),
+        }
+    }
+}
+
+fn print_bless_usage() {
+    println!();
+    println!("Possible options:");
+    println!("    `cargo bless`       : Bless only unspecified/new test cases");
+    println!("    `cargo bless --fails`: Bless everything, even failed test cases");
+}
 
 ///////////////////////////
 // TOP-LEVEL RUNNER CODE //
